refactor single and multi label metrics
