+++ mmeval/metrics/precision_recall_f1score.py
@@ -1,11 +1,12 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-
 import numpy as np
+import warnings
 from typing import (TYPE_CHECKING, Any, Dict, List, Optional, Sequence, Tuple,
                     Union, overload)
 
 from mmeval.core.base_metric import BaseMetric
 from mmeval.core.dispatcher import dispatch
+from mmeval.metrics.utils import MultiLabelMixin, format_data
 from mmeval.utils import try_import
 
 if TYPE_CHECKING:
@@ -20,11 +21,67 @@ else:
     flow = try_import('oneflow')
     of_F = try_import('oneflow.nn.functional')
 
-NUMPY_IMPL_HINTS = Tuple[Union[np.ndarray, np.number], np.number]
+NUMPY_IMPL_HINTS = Tuple[Union[np.ndarray, np.number], Union[np.ndarray,
+                                                             np.number]]
 TORCH_IMPL_HINTS = Tuple['torch.Tensor', 'torch.Tensor']
 ONEFLOW_IMPL_HINTS = Tuple['oneflow.Tensor', 'oneflow.Tensor']
 BUILTIN_IMPL_HINTS = Tuple[Union[int, Sequence[Union[int, float]]],
                            Union[int, Sequence[int]]]
+
+
+class PrecsionRecallF1score:
+    """Wrapper to get different task of PrecsionRecallF1score calculation, by
+    setting the ``task`` argument to either ``'singlelabel'`` or
+    ``multilabel``.
+
+    See the documentation of :mod:`SingleLabelPrecsionRecallF1score` and
+    :mod:`MultiLabelPrecsionRecallF1score` for the detailed usages and
+    examples.
+
+    Examples:
+        >>> import torch
+        >>> preds  = torch.tensor([2, 0, 1, 1])
+        >>> labels = torch.tensor([2, 1, 2, 0])
+        >>> metric = PrecsionRecallF1score(num_classes=3)
+        >>> metric(preds, labels)
+        {'precision': 33.3333, 'recall': 16.6667, 'f1-score': 22.2222}
+        >>> metric = PrecsionRecallF1score(
+                task="multilabel", average='micro', num_classes=3)
+        >>> metric(preds, labels)
+        {'precision_micro': 25.0, 'recall_micro': 25.0, 'f1-score_micro': 25.0}
+    """
+
+    def __new__(cls,
+                task: str = 'singlelabel',
+                num_classes: Optional[int] = None,
+                thrs: Union[float, Sequence[Optional[float]], None] = None,
+                topk: Optional[int] = None,
+                items: Sequence[str] = ('precision', 'recall', 'f1-score'),
+                average: Optional[str] = 'macro',
+                **kwargs):
+
+        if task == 'singlelabel':
+            return SingleLabelPrecsionRecallF1score(
+                num_classes=num_classes,
+                thrs=thrs,
+                items=items,
+                average=average,
+                **kwargs)
+        if task == 'multilabel':
+            assert isinstance(thrs, float) or thrs is None, \
+                "task `'multilabel'` only supports single threshold or None."
+            assert isinstance(num_classes, int), \
+                '`num_classes` is necessary for multi-label metrics.'
+            return MultiLabelPrecsionRecallF1score(
+                num_classes=num_classes,
+                thr=thrs,
+                topk=topk,
+                items=items,
+                average=average,
+                **kwargs)
+        raise ValueError(
+            f"Expected argument `task` to either be `'singlelabel'` or "
+            f"`'multilabel'` but got {task}")
 
 
 def _precision_recall_f1_support(pred_positive: Union[np.ndarray,
@@ -112,7 +169,7 @@ def _precision_recall_f1_support(pred_positive: Union[np.ndarray,
     return precision, recall, f1_score, support
 
 
-class SingleLabelMetric(BaseMetric):
+class SingleLabelPrecsionRecallF1score(BaseMetric):
     """A collection of metrics for single-label multi-class classification task
     based on confusion matrix.
 
@@ -153,8 +210,8 @@ class SingleLabelMetric(BaseMetric):
 
     Examples:
 
-        >>> from mmeval import SingleLabelMetric
-        >>> single_lable_metic = SingleLabelMetric(num_classes=4)
+        >>> from mmeval import SingleLabelPrecsionRecallF1score
+        >>> single_lable_metic = SingleLabelPrecsionRecallF1score(num_classes=4)
 
     Use NumPy implementation:
 
@@ -180,7 +237,7 @@ class SingleLabelMetric(BaseMetric):
             [0.3, 0.4, 0.2, 0.1],
             [0.0, 0.0, 0.1, 0.9]])
         >>> labels = np.asarray([0, 1, 2, 3])
-        >>> single_lable_metic = SingleLabelMetric(average='micro')
+        >>> single_lable_metic = SingleLabelPrecsionRecallF1score(average='micro')
         >>> single_lable_metic(preds, labels)
         {'precision_micro': 50.0, 'recall_micro': 50.0, 'f1-score_micro': 50.0} # noqa
 
@@ -194,10 +251,10 @@ class SingleLabelMetric(BaseMetric):
     """
 
     def __init__(self,
-                 thrs: Union[float, Sequence[Optional[float]], None] = 0.,
+                 num_classes: Optional[int] = None,
+                 thrs: Union[float, Sequence[Optional[float]], None] = None,
                  items: Sequence[str] = ('precision', 'recall', 'f1-score'),
                  average: Optional[str] = 'macro',
-                 num_classes: Optional[int] = None,
                  **kwargs) -> None:
         super().__init__(**kwargs)
 
@@ -271,11 +328,14 @@ class SingleLabelMetric(BaseMetric):
         for k, v in metrics.items():
 
             if self.average is None:
-                result_metrics[k + '_classwise'] = v.tolist()
+                _result = v.tolist()
+                result_metrics[k + '_classwise'] = [
+                    round(_r, 4) for _r in _result
+                ]
             elif self.average == 'micro':
-                result_metrics[k + f'_{self.average}'] = v.item()
+                result_metrics[k + f'_{self.average}'] = round(v.item(), 4)
             else:
-                result_metrics[k] = v.item()
+                result_metrics[k] = round(v.item(), 4)
 
         return result_metrics
 
@@ -469,4 +529,346 @@ class SingleLabelMetric(BaseMetric):
         predictions = [res[0] for res in results]
         labels = [res[1] for res in results]
         metric_results = self._compute_metric(predictions, labels)
+        return self._format_metric_results(metric_results)
+
+
+class MultiLabelPrecsionRecallF1score(MultiLabelMixin, BaseMetric):
+    """A collection of metrics for multi-label multi-class classification task
+    based on confusion matrix.
+
+    It includes precision, recall, f1-score and support.
+
+    Args:
+        num_classes (int): Number of classes. Needed for different inputs
+            as extra check.
+        thr (float, optional): Predictions with scores under the thresholds
+            are considered as negative. Defaults to None.
+        topk (int, optional): Predictions with the k-th highest scores are
+            considered as positive. Defaults to None.
+        items (Sequence[str]): The detailed metric items to evaluate. Here is
+            the available options:
+
+            - `"precision"`: The ratio tp / (tp + fp) where tp is the
+                number of true positives and fp the number of false
+                positives.
+            - `"recall"`: The ratio tp / (tp + fn) where tp is the number
+                of true positives and fn the number of false negatives.
+            - `"f1-score"`: The f1-score is the harmonic mean of the
+                precision and recall.
+            - `"support"`: The total number of positive of each category
+                in the target.
+
+            Defaults to ('precision', 'recall', 'f1-score').
+        average (str | None): The average method. It supports three average
+            modes:
+
+            - `"macro"`: Calculate metrics for each category, and calculate
+                the mean value over all categories.
+            - `"micro"`: Calculate metrics globally by counting the total
+                true positives, false negatives and false positives.
+            - `None`: Return scores of all categories.
+
+            Defaults to "macro".
+
+    .. note::
+        MultiLabelPrecsionRecallF1score supports different kinds of inputs. Such as:
+        1. Each sample has scores for every classes. (Only for predictions)
+        2. Each sample has one-hot indices for every classes.
+        3. Each sample has label-format indices.
+
+    Examples:
+
+        >>> from mmeval import MultiLabelPrecsionRecallF1score
+        >>> multi_lable_metic = MultiLabelPrecsionRecallF1score(num_classes=4)
+
+    Use Builtin implementation with raw indices:
+
+        >>> preds = [[0], [1], [2], [0, 3]]
+        >>> labels = [[0], [1, 2], [2], [3]]
+        >>> multi_lable_metic(preds, labels)
+        {'precision': 87.5, 'recall': 87.5, 'f1-score': 83.33}
+
+    Use Builtin implementation with one-hot indices:
+
+        >>> preds = [[1, 0, 0, 0],
+                      [0, 1, 0, 0],
+                      [0, 0, 1, 0],
+                      [1, 0, 0, 1]]
+        >>> labels = [[1, 0, 0, 0],
+                     [0, 1, 1, 0],
+                     [0, 0, 1, 0],
+                     [0, 0, 0, 1]]
+        >>> multi_lable_metic(preds, labels)
+        {'precision': 87.5, 'recall': 87.5, 'f1-score': 83.33}
+
+    Use Builtin implementation with scores:
+
+        >>> preds = [[0.9, 0.1, 0.2, 0.3],
+                      [0.1, 0.8, 0.1, 0.1],
+                      [0.4, 0.3, 0.7, 0.1],
+                      [0.8, 0.1, 0.1, 0.9]]
+        >>> labels = [[1, 0, 0, 0],
+                     [0, 1, 1, 0],
+                     [0, 0, 1, 0],
+                     [0, 0, 0, 1]]
+        >>> multi_lable_metic(preds, labels)
+        {'precision': 87.5, 'recall': 87.5, 'f1-score': 83.33}
+
+    Use NumPy implementation with raw indices:
+
+        >>> import numpy as np
+        >>> preds = [np.array([0]), np.array([1, 2]), np.array([2]), np.array([3])] # noqa
+        >>> labels = [np.array([0]), np.array([1]), np.array([2]), np.array([0, 3])] # noqa
+        >>> multi_lable_metic(preds, labels)
+        {'precision': 87.5, 'recall': 87.5, 'f1-score': 83.33}
+
+    Use PyTorch implementation:
+
+        >>> import torch
+        >>> preds = [torch.tensor([0]), torch.tensor([1, 2]), torch.tensor([2]), torch.tensor([3])] # noqa
+        >>> labels = [torch.tensor([0]), torch.tensor([1]), torch.tensor([2]), torch.tensor([0, 3])] # noqa
+        >>> multi_lable_metic(preds, labels)
+        {'precision': 87.5, 'recall': 87.5, 'f1-score': 83.33}
+
+    Computing with `micro` average mode with `topk=2`:
+
+        >>> preds = np.array([
+            [0.7, 0.1, 0.1, 0.1],
+            [0.1, 0.3, 0.4, 0.2],
+            [0.3, 0.4, 0.2, 0.1],
+            [0.0, 0.0, 0.1, 0.9]])
+        >>> labels = np.array([0, 1, 2, 3])
+        >>> multi_lable_metic = MultiLabelPrecsionRecallF1score(4, average='micro', topk=2)
+        >>> multi_lable_metic(preds, labels)
+        {'precision_top2_micro': 37.5, 'recall_top2_micro': 75.0, 'f1-score_top2_micro': 50.0} # noqa
+
+    Accumulate batch:
+
+        >>> for i in range(10):
+        ...     labels = torch.randint(0, 4, size=(100, ))
+        ...     predicts = torch.randint(0, 4, size=(100, ))
+        ...     multi_lable_metic.add(predicts, labels)
+        >>> multi_lable_metic.compute()  # doctest: +SKIP
+    """
+
+    def __init__(self,
+                 num_classes: int,
+                 thr: Optional[float] = None,
+                 topk: Optional[int] = None,
+                 items: Sequence[str] = ('precision', 'recall', 'f1-score'),
+                 average: Optional[str] = 'macro',
+                 **kwargs) -> None:
+        super().__init__(**kwargs)
+
+        if thr is None and topk is None:
+            thr = 0.5
+            warnings.warn('Neither thr nor k is given, set thr as 0.5 by '
+                          'default.')
+        elif thr is not None and topk is not None:
+            warnings.warn('Both thr and topk are given, '
+                          'use threshold in favor of top-k.')
+
+        self.thr = thr
+        self.topk = topk
+
+        for item in items:
+            assert item in ['precision', 'recall', 'f1-score', 'support'], \
+                f'The metric {item} is not supported by' \
+                ' `MultiLabelPrecsionRecallF1score`,' \
+                ' please specify from "precision", "recall", "f1-score" and ' \
+                '"support".'
+        self.items = tuple(items)
+
+        average_options = ['micro', 'macro', None]
+        assert average in average_options, 'Invalid `average` argument, ' \
+            f'please specify from {average_options}.'
+        self.average = average
+        self.num_classes = num_classes
+
+    def add(self, predictions: Sequence, labels: Sequence) -> None:  # type: ignore # yapf: disable # noqa: E501
+        """Add the intermediate results to `self._results`.
+
+        Args:
+            predictions (Sequence): Predictions from the model. It can be
+                labels (N, ), or scores of every class (N, C).
+            labels (Sequence): The ground truth labels. It should be (N, ).
+        """
+        for pred, label in zip(predictions, labels):
+            self._results.append((pred, label))
+
+    def _format_metric_results(self, results: List) -> Dict:
+        """Format the given metric results into a dictionary.
+
+        Args:
+            results (list): Results of precision, recall, f1 and support.
+
+        Returns:
+            dict: The formatted dictionary.
+        """
+        metrics = {}
+
+        def pack_results(precision, recall, f1_score, support):
+            single_metrics = {}
+            if 'precision' in self.items:
+                single_metrics['precision'] = precision
+            if 'recall' in self.items:
+                single_metrics['recall'] = recall
+            if 'f1-score' in self.items:
+                single_metrics['f1-score'] = f1_score
+            if 'support' in self.items:
+                single_metrics['support'] = support
+            return single_metrics
+
+        if self.thr:
+            suffix = '' if self.thr == 0.5 else f'_thr-{self.thr:.2f}'
+            for k, v in pack_results(*results).items():
+                metrics[k + suffix] = v
+        else:
+            for k, v in pack_results(*results).items():
+                metrics[k + f'_top{self.topk}'] = v
+
+        result_metrics = dict()
+        for k, v in metrics.items():
+
+            if self.average is None:
+                _result = v.tolist()
+                result_metrics[k + '_classwise'] = [
+                    round(_r, 4) for _r in _result
+                ]
+            elif self.average == 'micro':
+                result_metrics[k + f'_{self.average}'] = round(v.item(), 4)
+            else:
+                result_metrics[k] = round(v.item(), 4)
+
+        return result_metrics
+
+    @overload
+    @dispatch
+    def _compute_metric(self, predictions: Sequence['torch.Tensor'],
+                        labels: Sequence['torch.Tensor']) -> List:
+        """A PyTorch implementation that computes the metric."""
+
+        preds = format_data(predictions, self.num_classes,
+                            self._pred_is_onehot)
+        labels = format_data(labels, self.num_classes,
+                             self._label_is_onehot).long()
+
+        # cannot be raised in current implementation because
+        # `and` method will guarantee the equal length.
+        # However length check should remain somewhere.
+        assert preds.shape[0] == labels.shape[0], \
+            'Number of samples does not match between preds' \
+            f'({preds.shape[0]}) and labels ({labels.shape[0]}).'
+
+        if self.thr is not None:
+            # a label is predicted positive if larger than self.
+            # work for index as well
+            pos_inds = (preds >= self.thr).long()
+        else:
+            # top-k labels will be predicted positive for any example
+            _, topk_indices = preds.topk(self.topk)
+            pos_inds = torch.zeros_like(preds).scatter_(1, topk_indices, 1)
+            pos_inds = pos_inds.long()
+
+        return _precision_recall_f1_support(  # type: ignore
+            pos_inds, labels, self.average)
+
+    @overload  # type: ignore
+    @dispatch
+    def _compute_metric(  # type: ignore
+            self, predictions: Sequence['oneflow.Tensor'],
+            labels: Sequence['oneflow.Tensor']) -> List:
+        """A OneFlow implementation that computes the metric."""
+
+        preds = format_data(predictions, self.num_classes,
+                            self._pred_is_onehot)
+        labels = format_data(labels, self.num_classes,
+                             self._label_is_onehot).long()
+
+        # cannot be raised in current implementation because
+        # `and` method will guarantee the equal length.
+        # However length check should remain somewhere.
+        assert preds.shape[0] == labels.shape[0], \
+            'Number of samples does not match between preds' \
+            f'({preds.shape[0]}) and labels ({labels.shape[0]}).'
+
+        if self.thr is not None:
+            # a label is predicted positive if larger than self.
+            # work for index as well
+            pos_inds = (preds >= self.thr).long()
+        else:
+            # top-k labels will be predicted positive for any example
+            _, topk_indices = preds.topk(self.topk, dim=-1)
+            pos_inds = flow.zeros_like(preds).scatter_(1, topk_indices, 1)
+            pos_inds = pos_inds.long()
+
+        return _precision_recall_f1_support(  # type: ignore
+            pos_inds, labels, self.average)
+
+    @overload
+    @dispatch
+    def _compute_metric(self, preds: Sequence[Union[int,
+                                                    Sequence[Union[int,
+                                                                   float]]]],
+                        labels: Sequence[Union[int, Sequence[int]]]) -> List:
+        """A Builtin implementation that computes the metric."""
+
+        return self._compute_metric([np.array(pred) for pred in preds],
+                                    [np.array(target) for target in labels])
+
+    @dispatch
+    def _compute_metric(
+            self, preds: Sequence[Union[np.ndarray, np.number]],
+            labels: Sequence[Union[np.ndarray, np.number]]) -> List:
+        """A NumPy implementation that computes the metric."""
+
+        preds = format_data(preds, self.num_classes, self._pred_is_onehot)
+        labels = format_data(labels, self.num_classes,
+                             self._label_is_onehot).astype(np.int64)
+
+        # cannot be raised in current implementation because
+        # `and` method will guarantee the equal length.
+        # However length check should remain somewhere.
+        assert preds.shape[0] == labels.shape[0], \
+            'Number of samples does not match between preds' \
+            f'({preds.shape[0]}) and labels ({labels.shape[0]}).'
+
+        if self.thr is not None:
+            # a label is predicted positive if larger than self.
+            # work for index as well
+            pos_inds = (preds >= self.thr).astype(np.int64)
+        else:
+            # top-k labels will be predicted positive for any example
+            topk_indices = np.argpartition(
+                preds, -self.topk, axis=-1)[:, -self.topk:]  # type: ignore
+            pos_inds = np.zeros(preds.shape, dtype=np.int64)
+            np.put_along_axis(pos_inds, topk_indices, 1, axis=1)
+
+        return _precision_recall_f1_support(  # type: ignore
+            pos_inds, labels, self.average)
+
+    def compute_metric(
+        self, results: List[Union[NUMPY_IMPL_HINTS, TORCH_IMPL_HINTS,
+                                  ONEFLOW_IMPL_HINTS, BUILTIN_IMPL_HINTS]]
+    ) -> Dict[str, float]:
+        """Compute the metric.
+
+        Currently, there are 3 implementations of this method: NumPy and
+        PyTorch and OneFlow. Which implementation to use is determined by the
+        type of the calling parameters. e.g. `numpy.ndarray` or `torch.Tensor`
+        or `oneflow.Tensor`.
+        This method would be invoked in `BaseMetric.compute` after distributed
+        synchronization.
+
+        Args:
+            results (List[Union[NUMPY_IMPL_HINTS, TORCH_IMPL_HINTS,
+            ONEFLOW_IMPL_HINTS]]): A listof tuples that consisting the
+            prediction and label. This list has already been synced across all
+            ranks.
+        Returns:
+            Dict[str, float]: The computed metric.
+        """
+        preds = [res[0] for res in results]
+        labels = [res[1] for res in results]
+        metric_results = self._compute_metric(preds, labels)
         return self._format_metric_results(metric_results)
