+++ mmeval/metrics/utils/__init__.py
@@ -5,6 +5,7 @@ from .bbox_overlaps_rotated import (calculate_bboxes_area_rotated,
 from .grammar import get_n_gram, get_tokenizer, infer_language
 from .image_transforms import reorder_and_crop
 from .keypoint import calc_distances, distance_acc
+from .multi_label import MultiLabelMixin, format_data
 from .polygon import (poly2shapely, poly_intersection, poly_iou,
                       poly_make_valid, poly_union, polys2shapely)
 
@@ -13,5 +14,6 @@ __all__ = [
     'poly_make_valid', 'poly_iou', 'calc_distances', 'distance_acc',
     'calculate_overlaps', 'calculate_bboxes_area', 'reorder_and_crop',
     'calculate_bboxes_area_rotated', 'calculate_overlaps_rotated',
-    'get_n_gram', 'get_tokenizer', 'infer_language'
+    'get_n_gram', 'get_tokenizer', 'infer_language', 'MultiLabelMixin',
+    'format_data'
 ]
